[Refactor] Reduce ops/ file count: merge, rename, and reorganize
- Merge avg_pool1d/2d/3d.py → pool.py
- Merge conv1d/2d/3d.py → convolution.py
- Merge engram_fwd.py + engram_bwd.py → engram.py
- Drop _fwd suffix: da_cumsum, ssd_chunk_scan, ssd_chunk_state,
  ssd_state_passing, gqa_sliding_window, gqa_sliding_window_varlen
- Move 13 attention ops into ops/attention/ subpackage
- Update all imports in __init__.py, tests, and benchmarks

No functional changes. All public API symbols unchanged.
