+++ tileops/ops/__init__.py
@@ -1,18 +1,25 @@
-from .avg_pool1d import AvgPool1dOp
-from .avg_pool2d import AvgPool2dOp
-from .avg_pool3d import AvgPool3dOp
-from .conv1d import Conv1dBiasFwdOp, Conv1dFwdOp
-from .conv2d import Conv2dOp
-from .conv3d import Conv3dOp
-from .da_cumsum_fwd import DaCumsumFwdOp
-from .deepseek_dsa_decode import DeepSeekSparseAttentionDecodeWithKVCacheFwdOp
-from .deepseek_mla_decode import MultiHeadLatentAttentionDecodeWithKVCacheFwdOp
-from .deepseek_nsa import (
+from .attention import (
+    DeepSeekSparseAttentionDecodeWithKVCacheFwdOp,
+    GqaSlidingWindowFwdOp,
+    GqaSlidingWindowVarlenFwdOp,
+    GroupedQueryAttentionBwdOp,
+    GroupedQueryAttentionDecodePagedWithKVCacheFwdOp,
+    GroupedQueryAttentionDecodeWithKVCacheFwdOp,
+    GroupedQueryAttentionFwdOp,
     MeanPoolingForwardOp,
+    MHCPostOp,
+    MHCPreOp,
+    MultiHeadAttentionBwdOp,
+    MultiHeadAttentionDecodePagedWithKVCacheFwdOp,
+    MultiHeadAttentionDecodeWithKVCacheFwdOp,
+    MultiHeadAttentionFwdOp,
+    MultiHeadLatentAttentionDecodeWithKVCacheFwdOp,
     NSACmpFwdVarlenOp,
     NSAFwdVarlenOp,
     NSATopkVarlenOp,
 )
+from .convolution import Conv1dBiasFwdOp, Conv1dFwdOp, Conv2dOp, Conv3dOp
+from .da_cumsum import DaCumsumFwdOp
 from .deltanet_chunkwise import DeltaNetBwdOp, DeltaNetFwdOp, DeltaNetOp
 from .deltanet_recurrence import DeltaNetDecodeOp
 from .dropout import DropoutOp
@@ -25,17 +32,7 @@ from .gated_deltanet_recurrence import GatedDeltaNetDecodeOp
 from .gemm import GemmOp
 from .gla_chunkwise import GLABwdOp, GLAFwdOp
 from .gla_recurrence import GLADecodeOp
-from .gqa import GroupedQueryAttentionBwdOp, GroupedQueryAttentionFwdOp
-from .gqa_decode import GroupedQueryAttentionDecodeWithKVCacheFwdOp
-from .gqa_decode_paged import GroupedQueryAttentionDecodePagedWithKVCacheFwdOp
-from .gqa_sliding_window_fwd import GqaSlidingWindowFwdOp
-from .gqa_sliding_window_varlen_fwd import GqaSlidingWindowVarlenFwdOp
 from .grouped_gemm import GroupedGemmOp
-from .mha import MultiHeadAttentionBwdOp, MultiHeadAttentionFwdOp
-from .mha_decode import MultiHeadAttentionDecodeWithKVCacheFwdOp
-from .mha_decode_paged import MultiHeadAttentionDecodePagedWithKVCacheFwdOp
-from .mhc_post import MHCPostOp
-from .mhc_pre import MHCPreOp
 from .moe import MoePermuteAlignFwdOp
 from .norm import (
     AdaLayerNormFwdOp,
@@ -50,6 +47,7 @@ from .norm import (
     RMSNormFwdOp,
 )
 from .op import Op
+from .pool import AvgPool1dOp, AvgPool2dOp, AvgPool3dOp
 
 # --- Reduction ops (uncomment as sub-category PRs land) ---
 from .reduction import (
@@ -84,10 +82,10 @@ from .rope import (
     RopeNonNeoxOp,
     RopeYarnOp,
 )
-from .ssd_chunk_scan_fwd import SsdChunkScanFwdOp
-from .ssd_chunk_state_fwd import SsdChunkStateFwdOp
+from .ssd_chunk_scan import SsdChunkScanFwdOp
+from .ssd_chunk_state import SsdChunkStateFwdOp
 from .ssd_decode import SsdDecodeOp
-from .ssd_state_passing_fwd import SsdStatePassingFwdOp
+from .ssd_state_passing import SsdStatePassingFwdOp
 from .topk_selector import TopkSelectorOp
 
 __all__ = [
